% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstract-bmi.R
\docType{data}
\name{AbstractBmi}
\alias{AbstractBmi}
\title{Abstract class defining the basic model interface}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
AbstractBmi
}
\description{
Documentation on each function can be found at https://bmi.readthedocs.io/
}
\keyword{datasets}
