\name{bmi-package}
\alias{bmi-package}
\alias{bmi}
\docType{package}
\title{
\packageTitle{bmi}
}
\description{
\packageDescription{bmi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bmi}
\packageIndices{bmi}
}
\author{
\packageAuthor{bmi}

Maintainer: \packageMaintainer{bmi}
}
\keyword{ package }

