\name{grpc4bmi-package}
\alias{grpc4bmi-package}
\alias{grpc4bmi}
\docType{package}
\title{
\packageTitle{grpc4bmi}
}
\description{
\packageDescription{grpc4bmi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{grpc4bmi}
\packageIndices{grpc4bmi}
}
\author{
\packageAuthor{grpc4bmi}

Maintainer: \packageMaintainer{grpc4bmi}
}
\keyword{ package }

