% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-server.R
\name{runServer}
\alias{runServer}
\title{Run a grpc server which wraps the bmi model on a port.
The model should be a subclass of AbstractBmi}
\usage{
runServer(model, port)
}
\arguments{
\item{model}{an object that subclasses AbstractBmi}

\item{port}{a tcp port number to run the server on}
}
\value{
a grpc server
}
\description{
Run a grpc server which wraps the bmi model on a port.
The model should be a subclass of AbstractBmi
}
