% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-server.R
\name{wrapModelinService}
\alias{wrapModelinService}
\title{Wraps a model in a grpc service}
\usage{
wrapModelinService(model)
}
\arguments{
\item{model}{a sublcass of AbstractBmi}
}
\value{
a grpc service
}
\description{
Wraps a model in a grpc service
}
